Refactor into separate modules for pre processing and zoning
